Query the server logs

Query the server logs

vos_log(p = NA, collapse = NULL, just_errors = FALSE)

Arguments

p
a process object, returned by vos_process() or vos_start(). (will be restored from cache if not provided)

collapse
an optional character string to separate the lines in a single character string.

just_errors
logical, default FALSE. Set to TRUE to return just the lines that contain the term "error", which can be useful in debugging or validating bulk imports.

Value

Virtuoso logs as a character vector.

See also

Examples

if(has_virtuoso())
  vos_log()

#>  [1] ""                                                                                                          
#>  [2] "\t\tTue Jul 13 2021"                                                                                       
#>  [3] "10:53:59 { Loading plugin 1: Type `plain', file `wikiv' in `/usr/lib/virtuoso-opensource-6.1/hosting'"     
#>  [4] "10:53:59   WikiV version 0.6 from OpenLink Software"                                                       
#>  [5] "10:53:59   Support functions for WikiV collaboration tool"                                                 
#>  [6] "10:53:59   SUCCESS plugin 1: loaded from /usr/lib/virtuoso-opensource-6.1/hosting/wikiv.so }"              
#>  [7] "10:53:59 { Loading plugin 2: Type `plain', file `mediawiki' in `/usr/lib/virtuoso-opensource-6.1/hosting'" 
#>  [8] "10:53:59   MediaWiki version 0.1 from OpenLink Software"                                                   
#>  [9] "10:53:59   Support functions for MediaWiki collaboration tool"                                             
#> [10] "10:53:59   SUCCESS plugin 2: loaded from /usr/lib/virtuoso-opensource-6.1/hosting/mediawiki.so }"          
#> [11] "10:53:59 { Loading plugin 3: Type `plain', file `creolewiki' in `/usr/lib/virtuoso-opensource-6.1/hosting'"
#> [12] "10:53:59   CreoleWiki version 0.1 from OpenLink Software"                                                  
#> [13] "10:53:59   Support functions for CreoleWiki collaboration tool"                                            
#> [14] "10:53:59   SUCCESS plugin 3: loaded from /usr/lib/virtuoso-opensource-6.1/hosting/creolewiki.so }"         
#> [15] "10:53:59 OpenLink Virtuoso Universal Server"                                                               
#> [16] "10:53:59 Version 06.01.3127-pthreads for Linux as of Jan 14 2019"                                          
#> [17] "10:53:59 uses parts of OpenSSL, PCRE, Html Tidy"                                                           
#> [18] "10:54:00 SQL Optimizer enabled (max 1000 layouts)"                                                         
#> [19] "10:54:01 Compiler unit is timed at 0.005091 msec"                                                          
#> [20] "10:54:03 built-in procedure \"repl_undot_name\" overruled by the RDBMS"                                    
#> [21] "10:54:03 built-in procedure \"REPL_FQNAME\" overruled by the RDBMS"                                        
#> [22] "10:54:03 built-in procedure \"REPL_COLTYPE_PS\" overruled by the RDBMS"                                    
#> [23] "10:54:03 built-in procedure \"REPL_COLTYPE\" overruled by the RDBMS"                                       
#> [24] "10:54:06 Checkpoint started"                                                                               
#> [25] "10:54:06 Roll forward started"                                                                             
#> [26] "10:54:06 Roll forward complete"                                                                            
#> [27] "10:54:06 Checkpoint started"                                                                               
#> [28] "10:54:06 Checkpoint finished, log reused"                                                                  
#> [29] "10:54:06 Checkpoint started"                                                                               
#> [30] "10:54:06 Checkpoint finished, log reused"                                                                  
#> [31] "10:54:06 PL LOG: Can't get list of vad packages in /usr/share/virtuoso-opensource-6.1/vad/"                
#> [32] "10:54:06 Checkpoint started"                                                                               
#> [33] "10:54:07 Checkpoint finished, log reused"                                                                  
#> [34] "10:54:07 HTTP/WebDAV server online at 8890"                                                                
#> [35] "10:54:07 Server online at 1111 (pid 1053)"