Run a SPARQL query

Run a SPARQL query

vos_query(con, query)

Arguments

con
a ODBC connection to Virtuoso, from vos_connect()

query
a SPARQL query statement

Value

a data.frame containing the results of the query

Details

SPARQL is a graph query language similar in syntax SQL, but allows the use of variables to walk through graph nodes.

References

https://en.wikipedia.org/wiki/SPARQL

https://docs.ropensci.org/rdflib/articles/rdf_intro.html

See also

Examples

vos_status()

#> latest log entry: 10:54:07 Server online at 1111 (pid 1053)

#> [1] "sleeping"

# \donttest{
if(has_virtuoso()){
vos_start()
con <- vos_connect()

# show first 4 triples in the database
DBI::dbGetQuery(con, "SPARQL SELECT * WHERE { ?s ?p ?o } LIMIT 4")
}

#> Virtuoso is already running with pid: 1053

#>                                                                              s
#> 1                   http://www.openlinksw.com/virtrdf-data-formats#default-iid
#> 2          http://www.openlinksw.com/virtrdf-data-formats#default-iid-nullable
#> 3          http://www.openlinksw.com/virtrdf-data-formats#default-iid-nonblank
#> 4 http://www.openlinksw.com/virtrdf-data-formats#default-iid-nonblank-nullable
#>                                                 p
#> 1 http://www.w3.org/1999/02/22-rdf-syntax-ns#type
#> 2 http://www.w3.org/1999/02/22-rdf-syntax-ns#type
#> 3 http://www.w3.org/1999/02/22-rdf-syntax-ns#type
#> 4 http://www.w3.org/1999/02/22-rdf-syntax-ns#type
#>                                                         o
#> 1 http://www.openlinksw.com/schemas/virtrdf#QuadMapFormat
#> 2 http://www.openlinksw.com/schemas/virtrdf#QuadMapFormat
#> 3 http://www.openlinksw.com/schemas/virtrdf#QuadMapFormat
#> 4 http://www.openlinksw.com/schemas/virtrdf#QuadMapFormat

# }